Fix routing bug on exsiting session page
There was a bug where if you did not select an option it would finish
the claim in progress.

Now if you do not pick an option it will show a flash error message at
the top of the page. We decided that due to the expected number of users
that would see this page that it was not worth spending the additional
time to sort out a more comprehensive error message.