Add Rake task to export data needed for school check email
Each month, a service operator sends a verification email to claimants'
schools, asking them to confirm some details of the claim.

In order to do the mail merge for this email, the service operator needs
a developer to provide them with an export of some of the data from the
submitted claims for which the email has not yet been sent. Until now,
we have been doing this as an ad-hoc process where Tekin would run
something in the Rails console.

Since our future plans are to either continue running this monthly
export, or to automate the subject check email, it makes sense to move
some of this ad-hoc process into the codebase.

This commit adds a Rake task that we can run to do the export. It takes
a list of claim references for which the email has already been sent
(this will be provided by the service operator each time a developer
needs to run this task), and prints a CSV to the standard output.

Since AFAIK we cannot upload / download files to / from the place where
the Azure console is run, here's my anticipated workflow:

1. Take list of claim references from the service operator, and do some
local munging to create a comma separated list. Copy it to clipboard.

2. Connect to Azure console and run this task as
`rails export:school_check_email CLAIM_REFERENCES_TO_EXCLUDE='<paste the comma separated list>'`.

3. Copy the output of this task using terminal UI and save it to a
CSV file on local machine.

4. Create a Google Sheets spreadsheet with access controls that only
allow the service operator to view it. Import the CSV output into this
spreadsheet and share spreadsheet with service operator.

5. Delete the CSV file from local machine.