JS-based school search handles session timeouts
This resolves a bug where a user with an expired session performing
schools searches would result in an exception and not return any
results.

An `ActionController::UrlGenerationError` exception was being raised by
`timeout_claim_path(current_policy_routing_name)` because the routing
for schools search doesn't set a `params [:policy]` resulting in
`current_policy_routing_name` returning `nil`. The way we get around
this is to make `SchoolSearchController` `PartOfClaimJourney`, which
overrides `current_policy_routing_name` to load the policy routing name
from the session if it isn't present in the params.

This means we can now generate a valid timeout path for redirection, but
we still need to update the timeout code as school search performs AJAX
requests and expect JSON response. So we return an alternative JSON
payload for a timeout and update the JS code to handle these responses.

This should also resolve some of the
ActionController::InvalidAuthenticityToken exceptions we're seeing as
`PartOfClaimJourney` sets caching headers that reduce the browser
caching that often results in those getting raised.