Extend session cookie to work around CSRF issue
Some browsers have a behaviour where although they will delete a session
cookie when the app is shutdown, they will still serve a cached version
of the page on relaunch. This causes issues with CSRF protection as it
relies on the token embedded in the page/form matching the token in the
session cookie. So a user that then submits the form will get a 422
error (`ActionController::InvalidAuthenticityToken`) because their is
nothing in the session. We get around this by giving the session cookie
a life beyond the browser session. We can do this without risk of
exposing user data as we still explicitly expire a active claim and
admin sessions based on the `last_seen_at` timestamp.

See https://github.com/rails/rails/issues/21948 for further details on
the issue.