Replace checked and check_result with single value
In Geckoboard, if a claim is checked or not and the outcome of that
check are stored as two separate values. As we are limited to 10 fields
in a dataset, these must be combined to allow the addition of the award
amount.

Whilst this does mean reworking some of the dashboard panels, it does
not prevent us from displaying the same information.