Session refresh doesn't try redirect for timeout
We've seen requests come in for this endpoint with expired claim
sessions, resulting in the followig exception being raised:

  ActionController::UrlGenerationError: No route matches
    {:action=>"timeout", :controller=>"claims", :policy=>nil},
    possible unmatched constraints: [:policy]

This happened because the endpoint isn't namespaced to a policy. This is
definitely an edge-case as the only place this endpoint is hit from
during normal use is from the JS-based timeout dialogue that pops
up when a session is due to expire.

Because this endpoint is hit for both admin and claim sessions we should
stop relying on the before actions on `BasePublicController` and instead
explicitly clear sessions that have expired to prevent expired sessions
from being restarted.