Identify claimants by teacher reference instead of NI number
Because of the restrictions of payroll, we can only make one payment to
an individual in a given month. Therefor if a user claims for more than
one policy, we need a way to combine their claims into a single payment.
Previously this was done by matching claims with the same
National Insurance number (NiNo), this commit switches that matching to
be based on the claimant's teacher reference number (TRN):

- Historically, users enter their TRN more reliably than NiNo so there
  are few with errors, meaning fewer claims to have to correct
- There is no technically reason for us to be check the NiNo on claims
  during claim checking except because of this matching. If we switch to
  use the TRN instead, claim checkers can ignore the NiNo during claim
  checking, giving them one less thing to worry about (HMRC will deal
  with any errors in NiNo)
- We need to guarantee that the TRN is correct anyway as that will be
  the only identifiable field we will maintain for paid claims once the
  claim window is closed