Reduce the minimum length of the search term to three characters
Originally we had a minimum of of four characters as a way to limit the
amount of results shown to the user. Now that we have a hard limit of
50 results, it feels less important to be so strict with the minimum
length.

We felt that requiring a user to type four characters of their school's
postcode (which they might not fully remember) might be a bit much. By
having the minimum as three characters it would allow for greater
flexibility. For example, finding a school with a postcode M1 2WD, would
only require the user to remember M12 and not worry about the specific
postcode sector.