Make sure Logstash entries are always tagged with environment name
We want to make sure that the named_tag.environment field is included in
all log entries sent by our application. It’s the only way for us to
distinguish between log entries from local development (if enabled),
development or test. This is because log entries from those environments
all go into the same Logit stack.

Until now we have relied on config.log_tags, but this only affects
request logging. This means that other messages such as custom log
messages, or Active Job's automatic start / end messages, do not get
tagged properly. So, we set up a custom SemanticLogger formatter which
appends the tag. This means that all calls that use the Rails logger now
send this field.

I’ve removed the fallback to sending a value of “unspecified” - I cannot
think of a legitimate time when the ENVIRONMENT_NAME environment
variable would not be set.