Simplify JS-based school search code
This reverts changes made in 4b39a309 to try and resolve an issue where
AJAX requests for a school search with an expired/non-existent session
were resulting in exceptions being raised. Instead of making school
search part of the claim journey, we go the other way and make the
school search endpoint entirely open and not beholden to the current
session. This should get rid of the issues arising from the
`end_expired_claim_sessions` and `send_unstarted_claiments_to_the_start`
filters failing when the policy can't be decerned from either the
routing or an active session.