Check-your-answers copes with blank date-of-birth
This makes it so that visiting the check-your-answers page without
having fully completed a claim will not result in an exception.

It's not something that should happen under normal usage, but we have
seen at least one exception raised by a user visiting the
check-your-answers page with a claim that doesn't have a value for
date_of_birth. This updates the helper that displays the answer for the
question such that it won't raise an exception if the value is nil.