Allow payments to be removed from a Payroll Run
This shows a list of payments for a Payroll run, and, if the payroll run has not
yet been confirmed, a link to a page that allows a user to delete the payment.
If they deleted the payment, they see a confirmation screen with the
reference(s) of the claim(s) associated with that payment and a prompt
letting them know what the next steps are.