Remove `verify_response` from a Claim
This was introduced when we first started integrating with Verify, as we didn't
have a huge amount of confidence in the data it would return, so kept the raw
response against a claim too. It has come in handy in the past, but, as the
Verify integration is more or less stable, it makes less sense (and is a big risk)
for us to be storing this much PII, so we should remove it.