Add buttons to continue on from each check
At the moment these buttons don't do anything apart from going on to
the next check if there is one or the claim decision page if there are
no more checks. They are just hard-coded for now.

In the future we intend to mark these checks as complete in the
database.