Add academic_year to claims
Includes a data migration to seed the existing claims with the current
academic year.

This attribute will want a presence validation, but we'll have to wait
until we've shipped this code and run the data migration (plus the
future code that will set it for new claims), otherwise existing in-
progress claims will be returned as invalid until the data migration has
been run. Instead we'll add the validation in as a separate PR once this
has all shipped.