Display qualification and employment information
So that these new checks are useful to service operators its important
to display all the information to do with the qualification check,
including the ITT specialism etc. We only show the type of degree if
their ITT subject was something other than Maths or Physics.

It's also worth mentioning that we only show information that claims
checkers can actually check. ie. we collect some information along the
journey to help the claimant arrive at certain answers but the
information doesn't need checking by a claims checker.

Eventually we want to get to a point where these checks are automatic
and are sourced from the Database of Qualified Teachers.

I've hardcoded the date string "6 April 2018 to 5 April 2019"
for now, there's some other work going on to allow for multiple claim
years so it probably makes sense to leave it hardcoded and tackle it
later.