Academic year options roll over every autumn term
The new academic year starts the first week of September. To be keep
things simple we'll use September 1st as the point at which we roll over
to the next academic year. By then we should be long passed the point
at which it would make any sense to accept an application for the
previous academic year.

We also update the feature spec so that it will continue to work in
future academic years.