Copy all files when building the test Docker image
This is only for the test target and ensures that all the files that
could be linted locally are also linted on CI. We need to be mindful
of files that get added to the project, if they are secrets or
superfluous we should add them to the .dockerignore file.

This change doesn't increase the size of the Docker image by much but it
will increase build times slightly because of the way that Docker
caches layers of a Dockerfile, whenever a file is changed it will need
to redo that copy step.