Fix empty table header on view claims screen
The <th> (table headers) tag helps associate table cells with the
correct row/column headers. Having an empty <th> may result in cells
with missing or correct information.

If a cell needs to be an empty header then it should be a <td> instead.

This is needed for 1.3.1 Info and Relationships (Level A)