Make M&P academic year current in fixtures
This makes it easier to make small iterative changes to vary the QTS
question based on the configured academic year.

Having the current_academic_year set to the future was useful to show
that it could be different, but it makes it hard to iterate towards
having the app vary the QTS question based on the configured value. This
is beause the current year's values are hardcoded all over the shop,
including in the specs. Setting it to the current years means small
changes can be made whilst keeping the specs green. The idea being once
all the work has been done to remove the hardcoding, we should be able
to change this back to be a future year and if all is well, the specs
will stay green.