Persist the check to the database
In the next commit I'll hide the submit button when a check has been
completed so I don't think it necessary to add any error messages for
attempting to create an already existing check. The user shouldn't
get in a state where they will be able to carry out the action a second
time.

Introduce a method `next_check_path` on the controller to preserve the
green feature specs. In the future we'll probably want to extract this
out to it's own policy specific class as the checks for each policy
won't be exactly the same.