Configured academic year changes the QTS question
Now the QTS question and the ineligible screen will change the options
and the playback to the user based the currently configured academic
year for each policy.

This introduces a new policy-agnostic set of locale keys for QTS
question answers. Introducing these new keys makes it possible to move
to the new dynamic QTS year answers iteratively whilst keeping the test
suite green for the parts of the service that haven't been updated yet
(i.e the admin area).