Don’t pad formatted hours with whitespace
This changes the result of I18n.format(Time.current) from e.g.

"20 February 2020  3:11pm"

to

"20 February 2020 3:11pm"

The repeated whitespace has no effect on the output when rendered in a
web browser. However, it makes it difficult to check for the presence of
a formatted time when writing a feature spec, since the have_content
matcher will not detect a string with repeated whitespace.

The hyphen is explained in the documentation for Time#strftime.