Make AcademicYear a custom ActiveRecord type
This saves us having to manually marshal the string representation into
our value object each time we want to display it or perform operations
on it.

We've had to override the AcademicYear#hash method primarily to keep the
RSpec change matcher happy in the PiiScrubber specs that compare the
hash of attributes, but it will also make the type behave more
consistently if it's used as key in a Hash.