Update AdminChecksPresenters to vary the QTS year
The value that a claim checker sees when reviewing the qualification
details should change depending on the year the claim was made, based on
the rules for determining the first eligible claim year for a given
policy.

We've baked in an assumption here that this presenter will only ever be
used to display an eligible claim, and therefore we can assume the
locale to look up for the text is always going to be
"answers.qts_award_years.on_or_after_cut_off_date".