Display the status of a completed check
We display the name of the person who created the check and the time
they did it. We hide the "Complete X check and continue" button to
prevent another person overriding the check.

In the future we may want to add multiple states to checks such as
"failed" and reintroduce the submit button, but for now we just want
to mark the check as "completed".