Create Amendment model
We are going to allow service operators to amend certain attributes of a
claim. This commit introduces the new model - the frontend functionality
will follow in the next commits.

Each time they amend a claim, we will create an Amendment record. This
contains the details of which attributes were changed, and their old and
new values. It also provides a notes field where the service operator
must explain why they are making the amendment.

Because the amendment may contain personal data, we will need to remove
this data after a certain amount of time, the same as we do with claims.
So, we include a personal_data_removed_at timestamp.

I’m using the term “personal data” instead of PII because that’s what
GDPR calls it, and PII is considered a deprecated term according to
Wikipedia [1]. We should update the existing code and claims table to
use this naming. I will make a separate story for this.

[1] https://en.wikipedia.org/wiki/Personal_data