Add stylesheet for Timeline component from HMCTS Design System
We’re going to use this to display a timeline of a claim’s amendments.

Taken from src/components/hmcts-timeline/_timeline.scss in
https://github.com/hmcts/frontend, revision 0756daa.

Rich discovered this component on the GOV.UK Design System backlog:
https://github.com/alphagov/govuk-design-system-backlog/issues/105

I’ve modified the stylesheet to remove the hmcts-timeline__documents,
hmcts-timeline__document-item and hmcts-timeline__document-link rules.
This is because we won’t need them in our design, and because they refer
to the non-existent #{$hmcts-images-path}icon-document.svg which causes
SASS compilation to fail.

I’ve also had to run `rake prettier:fix`, which modifies this
stylesheet, to keep the build green.