A claim can only be amended if it is not payrolled or rejected
We need to be able to amend a claim up until the point that we tell
Cantium about it. After this, if there is a problem, we would remove the
claim from the payroll run.

Given our current criteria for removing personal data from a claim, this
implies that a claim cannot be amended if its personal data has been
removed. But I thought it safest to add an explicit check in case these
criteria change.