A service operator can amend a claim
This adds an "Amend claim" button to the admin/claims#show and admin
checks pages, which takes the service operator to a form where they can
enter new values for some of the claim’s attributes. When they submit
the form, we update the claim and create a new amendment record.

We had previously decided that we don’t want to allow service operators
to amend a claim, but we have changed our mind about this, now that we
have a better idea of some of the use cases, which happen quite
frequently and currently require a developer to make a change in the
database:

- add payroll gender when it is not returned by Verify and the claimant
doesn’t know it
- correct a teacher reference number
- resolve payment issues with Cantium
- resolve inconsistencies in payment details between multiple claims
from the same claimant in the same month
- correct a student loan plan

This change means that we’re exposing claimants’ bank details to service
operators. Kristi confirmed that this is okay.

We introduce the Claim.amend method, which is responsible for updating
the claim, and then taking the changed values and using them to create
an amendment. Some things that made this require a bit of thought
technically:

- handling attributes which are modified in a before_save hook
- figuring out how to represent validation errors on the form
- making sure that we handle mass assignment of dates

Some issues to note:

- In a show and tell of this feature, Mark pointed out that putting all of
the fields on a single page is a risky antipattern that could lead to
accidental edits. We agreed that we will allow it for a first iteration
of the feature.

- We reuse the same validation error messages as those which are used for
the claimant journey. This means that the service operator might see
some validation error messages that don’t quite make sense to them, such
as “Enter your date of birth”. I spoke to Tekin about this and we
decided it wasn’t worth the effort to sort this out.

- There is some messaging that explains that a claim cannot be amended
because it is missing payroll gender or has inconsistent payment details
with another claim from the claimant. These messages suggest that there
is nothing that the service operator can do to resolve the situation.
This is no longer true. However, we decided that we don’t want to
rewrite these messages at the moment.

Some styling help from Rich.

Co-authored-by: Rich Cooley <rich@dxw.com>