Add City Technology Colleges to state funded schools
We determine state funded schools by two pieces of information: the
school type group and the school type. We explicitly rule out schools
that have a type group of independent as not being state funded.
However it turns out that City Technology Colleges are state funded and
have a type group of independent.

We have another card in the backlog to look at any other types of
state funded schools that we're missing from our rules.