Use Eligibilty#qts_award_year_answer where pos
DRY up the presenters and simplify the specs by using the newly created
Eligibilty#qts_award_year_answer method. Note we've changed the
behaviour slightly on the EligibilityAnswersPresenters – previously they
were basing the displayed year based on the policy's configured current
academic year, whereas now they are basing it on the academic year of
the claim. In real-world operation they will always be the same as the
configured academic year will not change except after a claim period has
ended and before the next claim period starts.

Using this in the admin presenters also means that the note about the
assumption on the value of qts_award_year no longer holds so can be
removed.