Service operators are taken to Maths and Physics tasks
This is a temporary step to encourage service operators to check Maths
and Physics claims using the new "checks" or "tasks" interface.

We've introduced the word "tasks" here to describe the new checks
journey, we may want to rename our references of checks to tasks so our
code matches our domain/UI.

Eventually this will be the default for all claims, but we're not quite
ready to check all information belonging to a Student Loans claim.

I've been explicit in our feature tests about the claim's policy
that gets checked, otherwise we rely on the value from factory making
our test pass. Eventually when this becomes the default we can remove
the restriction on the test setup.