Handle byte order mark when reading payment confirmation CSV
The Payment Confirmation Report that Cantium sent for February 2019
contains a byte order mark. We weren't stripping out these characters
so the CSV was failing to parse[1].

By passing an encoding of "BOM|UTF-8" to CSV#read we can handle when
a byte order mark is both present and absent.

[1]: https://bugs.ruby-lang.org/issues/15210