Redirect to a decision form after last check
As part of the changes to the new admin checks we want to show the
claim decision form after the last check has been completed.

There's currently no other way to get to this page, this will need
revisiting later after we've spoken to the Service/Interaction designer.

We currently have no warning messages on this new page so a service
operator won't know the reason a claim isn't approvable. The idea is
that these warnings will be converted in to tasks/checks that a service
operator can carry out, however until we get to that point we will need
to keep them. These changes will come in the next commit.