Display warning messages on decision form
There are two warnings preventing approval (missing payroll gender and
and multiple claims from the same person with different details) and
a warning about a manual identity check that requires extra action from
a service operator.

The idea is to move these warnings in to their own checks, but until we
get there it's important information for a service operator deciding on
a claim's outcome.

Because we're sharing the form with the claims#show page there's a bit
of duplication in the specs, I think that's ok as this change is just
temporary until we can repurpose the warnings in to checks.

The only way to get to the new decision form is by completing the last
check, I've reflected this in the feature spec.