User who uploaded confirmation report recorded as reference
We now record user information in a table in our database. This commit
updates adds a reference from the `payroll_run` table to the
`dfe_sign_in_users` and updates the code to reflect the new association.

We will need a data migration to migrate previous records over to the
new association.