Change the static error page feedback link to a mailto
We currently link to a Google Forms form which is specific to student
loans, but these pages are shared between both policies.

The GOV.UK Design System phase banner component [1] says that this link
is allowed to send an email, so change it to a mailto to the general
contact address that we give on the page.

[1] https://design-system.service.gov.uk/components/phase-banner/#add-a-feedback-link