Remove links from static error pages’ footers
Specifically:
- Contact us - not a great loss since the error pages already contain a
contact email address
- Cookies
- Terms and conditions
- Privacy Notice
- Accessibility Statement

These are all currently dead links, because we only have policy-specific
versions of those pages.

It’s better to have no links than to have dead links, so the immediate
thing we do is to remove them.

There remains an open question of whether this page should contain these
links - in both a regulatory and a user experience sense - and what they
should link to if so. I asked about this in dxw and cross-government
Slacks. Opinions are mixed. Some people said that the pages are part of
the service, so they should contain the links. Others said that they
don’t think it’s necessary. If we wanted to have the links, we could:

- link to menu pages which link to all of the policies’ versions of
these pages
- make these pages dynamic so that they show the correct page for the
policy you’re trying to claim for - some people think that this is risky
and that these error pages should always be static, though
- introduce generic versions of these error pages for a service called
“Claim additional payments for teaching”