Add CPU and memory alerts for worker container instance
Unfortunately CpuPercentage and MemoryPercentage metrics aren't
available for container instance monitoring. We have to use CpuUsage and
MemoryUsage instead, this means the threshold has to be a concrete
value rather than a proportion.

- CpuUsage is measured in millicores (1,000 per core)
- MemoryUsage is measured in bytes (1,000,000 per mb)

I've set the values as 800 millicores (80% of our 1 core) and
1200MB (80% of our 1500MB available memory).

We could parameterise these units in the container instance deploy and
pass them in to the app_alerts.json then use `[mul(value, 0.8)]` however
these values aren't likely to change so it feels like overkill.

The `app_alerts.json` template is used by the main app and the VSP
however we only want to create the worker alerts for the main app so
we need the condition based on the worker container instance ID.

https://docs.microsoft.com/bs-cyrl-ba/azure/azure-monitor/platform/metrics-supported#microsoftcontainerinstancecontainergroups