Allow claims to have multiple decisions
This lays the groundwork for decisions being undoable (setting the claim
back to an undecided status) and allowing a subsequent change of
decision whilst still retaining a complete history of decision changes.

The new `decision` method returns the latest decision on a `Claim`, and
is functionally identical to current behaviour.

There is one notable pair of changes to specs, which is the removal of
the spec ensuring payroll runs cannot be generated when there would be
two claims containing the same personal details but different bank
numbers. Reaching this point should not be possible, as claims with
matching details but different bank numbers cannot be approved. This
behaviour is now explicitly tested in `decision_spec.rb` to prevent
regressions.

The original test now fails where it previously passed since `:approved`
and `:rejected` traits in the claims factory now use `create` instead of
`build`. This change ensures tests pass where previously they would fail
due to reliance on the `decision` relation within a `Claim` returning
values even when unsaved.

Other changes to specs includes swapping `build` to `create` for
amendment specs, for the same underlying reason.