Expose Git commit hash in Rollbar
When investigating a previous error in Rollbar, I wanted to know which
version of the code was running when the error occurred. We can achieve
this by setting Rollbar’s `config.code_version`.

I think it makes sense to bake the Git commit hash into the Docker
image, since it provides metadata about which code is contained inside
the image and the context in which the image was built. I’ve put the
ENV instruction below the COPY instructions for caching reasons, since
the commit hash changes more frequently than the copied files.

I haven’t found a way to test this end to end. I’ve confirmed that the
`web` image built by CI, which gets pushed to Docker Hub, contains the
GIT_COMMIT_HASH environment variable, but I don't know if there's an
easy way to actually test that the exceptions triggered end up tagged
properly in Rollbar.