Expose Git commit hash in Rollbar
When investigating a previous error in Rollbar, I wanted to know which
version of the code was running when the error occurred. We can achieve
this by setting Rollbar’s `config.code_version`.

I think it makes sense to bake the Git commit hash into the Docker
image, since it provides metadata about which code is contained inside
the image and the context in which the image was built. I’ve put the
ENV instruction below the COPY instructions for caching reasons, since
the commit hash changes more frequently than the copied files.

I tested this by temporarily changing azure-pipelines.yml so that it
pushed the Docker image for this feature branch, and then updating my
local copy of bin/azure-deploy so that DOCKER_IMAGE_TAG pointed to this
image, and then deploying to the test environment. There, I started a
Rails console and called Rollbar.error("Lawrence testing"). I confirmed
that the error occurrence in Rollbar pointed to the correct commit.

For future reference, the Build.SourceVersion predefined variable is
listed here:
https://docs.microsoft.com/en-us/azure/devops/pipelines/build/variables?view=azure-devops&tabs=yaml