Change tags used on view claims for accessibility
As part of the accessibility pass it was noticed that the orange tags,
'tag--warning' had insufficient contrast. To get sufficient contrast the
orange background colour would need to be changed to a muddy colour
which would have impacted the colour coding on the SLA tags shown on
this page.

Also the class was being used as a warning to indicate which claims
had not gone through GOV.UK Verify, which personally seemed like it was
implying a problem where one did not exist.

A new tag modifier class was made and used, this highlights information
we feel is important to see at a glance.