Remove the `AZURE_BUILD_NUMBER` variable
We introduced this as a lightweight way of letting the app service know that
something had changed, and trigger a restart. We now don't need to do this,
as the change of the Docker Image is enough to trigger a restart, so having
this additional variable is just unnecessary bloat.